+++ common/src/main/java/com/dfsek/terra/world/generation/math/interpolation/NoiseChunkInterpolator.java
@@ -20,8 +20,8 @@ public class NoiseChunkInterpolator implements ChunkInterpolator {
             }
         }
 
-        for(byte x = 0; x < 4; x++) {
-            for(byte z = 0; z < 4; z++) {
+        for(int x = 0; x < 4; x++) {
+            for(int z = 0; z < 4; z++) {
                 for(int y = 0; y < 64; y++) {
                     interpGrid[x][y][z] = new Interpolator3(
                             noiseStorage[x][z][y],
